Import trickle 1.03

+++ bwstat.h
@@ -4,7 +4,7 @@
  * Copyright (c) 2003 Marius Aamodt Eriksen <marius@monkey.org>
  * All rights reserved.
  *
- * $Id: bwstat.h,v 1.3 2003/03/05 08:06:39 marius Exp $
+ * $Id: bwstat.h,v 1.4 2003/03/29 06:22:17 marius Exp $
  */
 
 #ifndef TRICKLE_BWSTAT
@@ -34,7 +34,7 @@ struct bwstat {
 int             bwstat_init(uint);
 struct bwstat  *bwstat_new(void);
 struct bwstat  *bwstat_free(struct bwstat *);
-void            bwstat_update(struct bwstat *, int, short);
+void            bwstat_update(struct bwstat *, size_t, short);
 struct timeval *bwstat_getdelay(struct bwstat *, size_t *, uint, short);
 struct bwstat  *bwstat_gettot(void);
